+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `onReceivedError` callback to capture and propagate `WebResourceError` objects via the `widget.onError` handler.
+- Added `onReceivedHttpError` callback to capture and propagate `WebResourceResponse` objects via the `widget.onError` handler.
+
 ## [1.0.5] - 2025-04-26
 
 - Added `portone_flutter_v2.podspec` for CocoaPods integration.
@@ -90,6 +97,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modified field type in PaymentRequest model (changed `isEscrow` to nullable)
 - Changed `paypal` to `convenienceStore` in the `PaymentPayMethod` enumeration
 
+[Unreleased]: https://github.com/AndrewDongminYoo/portone_flutter/compare/1.0.5...HEAD
 [1.0.5]: https://github.com/AndrewDongminYoo/portone_flutter/compare/1.0.4...1.0.5
 [1.0.4]: https://github.com/AndrewDongminYoo/portone_flutter/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/AndrewDongminYoo/portone_flutter/compare/1.0.2...1.0.3